make a /boot/GRIST_BOOT_KEY page for diagnosing configuration problems
This is a start at a page for diagnosing problems while setting up
Grist. Starting to add some diagnostics based on feedback in github
issues. We should make Grist installation easier! But when there is
a problem it should be easier to diagnose than it is now, and this
may help.