docs: better explain GRIST_DATA_DIR
This little variable has unexpected complexity: it defaults to being
relative to the working directory, which usually is the same as the
app directory.

However, in practice we instruct people to explicitly set this to
`/persist/docs`, so we should mention this common use case.

Some day I would like to simplify some of the logic around this
variable.

closes: #931