(core) Fix TypeORM failure when unremoving entities
Summary:
When `save` was called without any modifications to the
entity, TypeORM would throw an error. This could sometimes
occur if an entity was unremoved by two different clients with
the same application state. We now avoid calling `save` if a
document or workspace is already unremoved.

Test Plan: Manual.

Reviewers: jarek

Reviewed By: jarek

Differential Revision: https://phab.getgrist.com/D4365