move docker image back to node 18 / buster
This partially undos a recent upgrade because it results in
`gvisor` sandboxing not functioning. Hopefully this is temporary,
to give time to track down the cause, fix it, and extend CI testing
on candidate docker images to include `gvisor` sandboxing.

Upgrade: https://github.com/gristlabs/grist-core/pull/1255

Sorry for the breakage!