(core) Fix broken 2-way references tooltip
Summary:
A bit of code that logged telemetry when a tooltip was opened
wasn't updated to support bindable values. This only affected
the tooltips for 2-way references, and only when the telemetry
level was set to "full".

Test Plan:
Repro steps:

Set GRIST_TELEMETRY_LEVEL to full and restart your instance.
Try to open a 2-way reference tooltip - a circular reference error should be thrown.
Tested that with this fix, the error is no longer thrown, and the other kinds of tooltips (clickable, hoverable, etc.) still work as expected.

Reviewers: jarek

Reviewed By: jarek

Differential Revision: https://phab.getgrist.com/D4382