OIDC: overwrite callback URL in browser history on error
When we receive an OIDC callback, the normal behavior is to
finish the login process by redirecting the user to the
originally requested page. The `?next=` query parameter carries
that URL.

But when the OIDC callback results in an error (for example,
because the user's email is not verified), we render an error
page instead of redirecting the user to their destination.

This causes the original destination to be lost, and if the
user retries their login (either by refreshing the page or
clicking the button to attempt a new login) the new target
URL becomes the callback URL, which has no useful purpose.

To avoid this, we modify the error page to overwrite the
browser's current URL with the requested page's URL. That
way, whether the user refreshes the page or clicks the button,
the new attempt will have the same target URL as the first
one.

Co-authored-by: Florent <florent.git@zeteo.me>