Dockerfile: don't mix up Docker and debian versions of python 2.7
When the Python 3 and Python 2 collecting images
were separated, the `python:2.7-slim-buster` image started
being used to download a Python 2.7 binary.

Unfortunately the `apt-get install python2` command that
was necessary when the base image was `python:3.11-slim-buster`
was left in, causing two installations of Python 2.7 to
exist in the resulting `collector-py2` image.

This resulted in a somewhat broken `python2.7` in the final
image, that required a new configuration script in `/etc/python2.7`.
But that script wasn't exposed to the gvisor sandbox, causing
Python 2 to fail.

By starting from the `debian:buster-slim` image and
using only the Apt-installed `python2` as before, compatibility
with the existing gvisor sandbox is maintained without any
workaround.