(core) Remove unused code related to ancient sharing logic.
Summary:
Some code has been around related to old attempts at end-to-end
encrypted sharing. Those were never finished, have been unused, and are
unlikely to be picked up again in this form.

This also replaces the hand-rolled WorkCoordinator with async-mutex, which we
already use for the same purpose elsewhere.

Test Plan:
Only unused functionality is removed. Some is refactored and should have no
changes in behavior. Relying on existing tests to confirm that.

Reviewers: georgegevoian

Reviewed By: georgegevoian

Differential Revision: https://phab.getgrist.com/D4403