(themes) Implementation idea on new design tokens handling
Until now themes could only describe each component variable. This made
creating a theme difficult: we had to override all component variables
(hundreds of them) if we wanted to change one color shade used globally.

Here is an idea of a cssVars.designTokens prop that is a combination of
current `colors`, `vars` and `theme` props:

- all design tokens are overridable by a theme
- colors are now named in a more abstract way so that it makes sense for
a theme changing the primary color

Everything is explained in details in the cssVars.designTokens comments