make "yarn cli" command available as "docker run ... cli" for non-developers
There is a utility called "yarn cli" available if you have built
Grist yourself and have developer tools installed. It isn't a
particularly well thought out or carefully planned utility, but
it comes in handy, so this change makes it available via docker,
so people comfortable with containers but not building from
source code can use it. Example of use to dump the history of
actions embedded in a document:

```
docker run --rm -v $PWD:$PWD -it gristlabs/grist \
   cli sqlite query $PWD/docs/4gtUhAEGbGAdsGNc52k4H6.grist \
  --json "select * from _gristsys_ActionHistory"
```