(core) Fix bugs with attachment cells
Summary:
When uploading an attachment from the add row, the preview wasn't
showing until the row was added. The cause was a malformed URL
that included blank values for the cell query parameters.

Uploading an attachment within a card or card list by
clicking the inline icon would unintentionally open the attachment
editor due to unhandled event bubbling. Opening references via
the link icon was also affected; both are now fixed.

When deleting an attachment from a card or card list by hitting
Backspace/Delete, the attachment editor would open and nothing
would be deleted. This was caused by an inconsistency with how
DetailView handled the `clearValues` command; it now shares a
similar implementation to `GridView`, meaning that the contents
of selected fields are cleared, and the editor is not activated unless
the field contains a formula.

Test Plan: Browser tests.

Reviewers: jarek

Reviewed By: jarek

Subscribers: paulfitz

Differential Revision: https://phab.getgrist.com/D4406