Implement SCIM Groups and Roles endpoint
This change also introduce a "Team" and a "Role" type for Groups. "Role"
type is the historical type, it is the role granted to the user for a
resource. The "Team" type is a new type, it is meant to gather people
and grant them a role together for a resource.

The SCIM /Groups endpoints gives access to the "Teams" groups.
The SCIM /Roles endpoints gives a (limited) access to the "Roles"
groups.

For more information, please take a look at this Pr description:
https://github.com/gristlabs/grist-core/pull/1357#issue-2767744160

Co-authored-by: Grégoire Cutzach <gregoire@cutzach.com>