Adds initial support for external attachments using the Filesystem or MinIO (#1320)
When attachments are uploaded to Grist, they're stored in an SQLite table within the document. 

The result is that large files (multi-GB) substantially increase the size of the SQLite file, and cause performance problems (for example, with snapshots).

The current design / full scope of changes can be seen [in this Google document.](https://docs.google.com/document/d/1ST_DuR22llDyx4PVAMdBlHDz22L5QIJ4qIyLMOUOtZ8/edit#heading=h.d487w4iic6q).

This commit contains a minimal proof of concept for external attachments, with the following features:
- External attachment storage using either MinIO or the local filesystem.
- A "default attachment store" setting for documents, which sets the destination for newly uploaded attachments.
- Defaulting to existing behaviour. Without additional configuration, Grist's attachment storage behaviour is unchanged.
- MinIO attachment storage is automatically made available if it's configured for document storage, but not automatically used by documents.

No UI exists for configuring anything related to attachment storage at this time. To use a store, the API or an SQLite client needs using to set the `idOfDefaultAttachmentStore` document setting to the ID of the desired attachment store. Attachment store IDs are logged to the console.

Currently only a single store of each backend type (minio, filesystem) can exist, as they're automatically configured using the existing setting. 

Three significant components are added in this commit, with the rest of the commit being generally glue around them:
- `AttachmentFileManager` (in "AttachmentFileManager.ts") handles attachment fetching / saving logic on behalf of `ActiveDoc`
- `IAttachmentStore` is an instantiated store, that handles saving / loading attachments from a specific provider.
- `AttachmentStoreProvider` is used to resolve store IDs to stores, and is the main way of providing other components (e.g AttachmentFileManager) with access to store instances.

The available types of attachment store are defined in `coreCreator.ts`, under the name "attachmentStoreBackends". 

A few existing classes have been extended with streaming methods, to simplify interaction with the existing attachment code and to facilitate large files in the future without keeping copies in memory.